udpate cron jobs

+++ src/app/api/auth/delete-unverified-users/route.ts
@@ -7,7 +7,12 @@ const pool = createPool({
 
 export const dynamic = "force-dynamic";
 
-export async function GET() {
+export async function GET(request: Request) {
+  const authHeader = request.headers.get("authorization");
+  if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   const deleteResult = await pool.sql`
     DELETE FROM users
     WHERE email_verified = FALSE
